requirements.txt
pygame
python-chess
pyreason
networkx
